Adds basic driving and grabbing functions

+++ include/Lift15.hpp
@@ -8,7 +8,7 @@ public:
   Lift15(int llm, int lrm, int ulm, int urm, int am, int bgm, int fgm);
   ~Lift15();
   void moveMotorToHeight(int degrees);
-  void grab(int power);
+  void grab(double power);
 
 private:
   int heights[9];
